backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
google-cloud-translate==3.11.1
python-dotenv==1.0.0


